GO.db {GO.db}	R Documentation

Bioconductor annotation data package

Description

Welcome to the GO.db annotation Package. The purpose of this package is to provide detailed information about the latest version of the Gene Ontologies. This package is updated biannually.

Objects in this package are accessed using the select() interface. See ?select in the AnnotationDbi package for details.

See Also

AnnotationDb-class for use of keys(), columns() and select().

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.
columns(GO.db)

## Bimap interface:
## The 'old style' of interacting with these objects is manipulation as 
## bimaps. While this approach is still available we strongly encourage the
## use of select().
ls("package:GO.db")

[Package GO.db version 3.12.1 Index]