GOBPOFFSPRING {GO.db}	R Documentation

Annotation of GO Identifiers to their Biological Process Offspring

Description

This data set describes associations between GO molecular function (BP) terms and their offspring BP terms, based on the directed acyclic graph (DAG) defined by the Gene Ontology Consortium. The format is an R object mapping the GO BP terms to all offspring terms, where an offspring term is a more specific GO term that is preceded by the given GO term in the DAG (in other words, the children and all their children, etc.).

Details

Each GO BP term is mapped to a vector of offspring GO BP terms.

Biological process is defined as the broad biological goals, such as mitosis or purine metabolism, that are accomplished by ordered assemblies of molecular functions as defined by Gene Ontology Consortium.

Mappings were based on data provided by: Gene Ontology http://current.geneontology.org/ontology/go-basic.obo With a date stamp from the source of: 2020-09-10

References

http://www.geneontology.org/ and https://www.ncib.nlm.nih.gov/entrez/query.fcgi?db=gene

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
# Convert the object to a list
xx <- as.list(GOBPOFFSPRING)
# Remove GO IDs that do not have any offspring
xx <- xx[!is.na(xx)]
if(length(xx) > 0){
  # Get the offspring GO IDs for the first two elents of xx
  goids <- xx[1:2]
}

[Package GO.db version 3.12.1 Index]