GOBPPARENTS {GO.db}	R Documentation

Annotation of GO Identifiers to their Biological Process Parents

Description

This data set describes associations between GO molecular function (BP) terms and their direct parent BP terms, based on the directed acyclic graph (DAG) defined by the Gene Ontology Consortium. The format is an R object mapping the GO BP terms to all direct parent terms, where a direct parent term is a more general GO term that immediately precedes the given GO term in the DAG.

Details

Each GO BP term is mapped to a named vector of GO BP terms. The name associated with the parent term will be either isa, hasa or partof, where isa indicates that the child term is a more specific version of the parent, and hasa and partof indicate that the child term is a part of the parent. For example, a telomere is part of a chromosome.

Biological process is defined as the broad biological goals, such as mitosis or purine metabolism, that are accomplished by ordered assemblies of molecular functions as defined by Gene Ontology Consortium.

Mappings were based on data provided: Gene Ontology http://current.geneontology.org/ontology/go-basic.obo With a date stamp from the source of: 2020-09-10

References

http://www.geneontology.org/ and https://www.ncib.nlm.nih.gov/entrez/query.fcgi?db=gene

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
# Convert the object to a list
xx <- as.list(GOBPPARENTS)
# Remove GO IDs that do not have any parent
xx <- xx[!is.na(xx)]
if(length(xx) > 0){
    # Get the children GO IDs for the first elents of xx
    goids <- xx[[1]]
    # Find out the GO terms for the first parent goid
    GOID(GOTERM[[goids[1]]])
    Term(GOTERM[[goids[1]]])
    Synonym(GOTERM[[goids[1]]])
    Secondary(GOTERM[[goids[1]]])
    Definition(GOTERM[[goids[1]]])
    Ontology(GOTERM[[goids[1]]])
}

[Package GO.db version 3.12.1 Index]