GOOBSOLETE {GO.db}	R Documentation

Annotation of GO identifiers by terms defined by Gene Ontology Consortium and their status are obsolete

Description

This is an R object mapping GO identifiers to the specific terms in defined by Gene Ontology Consortium and their definition are obsolete

Details

All the obsolete GO terms that are collected in this index will no longer exist in other mapping objects.

Mappings were based on data provided by: Gene Ontology http://current.geneontology.org/ontology/go-basic.obo With a date stamp from the source of: 2020-09-10

References

https://www.ncib.nlm.nih.gov/LocusLink

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
# Convert the object to a list
xx <- as.list(GOTERM)
if(length(xx) > 0){
    # Get the TERMS for the first elent of xx
    GOID(xx[[1]])
    Ontology(xx[[1]])
}

[Package GO.db version 3.12.1 Index]