GOSYNONYM {GO.db}	R Documentation

Map from GO synonyms to GO terms

Description

GOSYNONYM is an R object that provides mapping from GO synonyms to GO terms

Details

[TODO: Put some details here]

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
x <- GOSYNONYM
sample(x, 3)
# GO ID "GO:0009435" has a lot of synonyms
GOTERM[["GO:0009435"]]
# GO ID "GO:0006736" is a synonym of GO ID "GO:0009435"
GOID(GOSYNONYM[["GO:0006736"]])

[Package GO.db version 3.12.1 Index]