GOTERM {GO.db}	R Documentation

Annotation of GO Identifiers to GO Terms

Description

This data set gives mappings between GO identifiers and their respective terms.

Details

Each GO identifier is mapped to a GOTerms object that has 6 slots: GOID: GO Identifier Term: The term for that GO id Synonym: Synonymous terms Secondary: Secondary terms that have been merged into this term Definition: Further definition of the GO term Ontology: One of MF - molecular function, BP - biological process, or CC - cellular component

All the obsolete GO terms are under the nodes "obsolete molecular function" (GO:0008369), "obsolete cellular component" (GO id GO:0008370), and "obsolete biological process" (GO:0008371). Each of these GO identifiers has a group of GO identifiers as their direct children with GO terms that were defined by GO but are deprecated in the current build. These deprecated GO terms were appended by "(obsolete)" when the data package was built.

Mappings were based on data provided by: Gene Ontology http://current.geneontology.org/ontology/go-basic.obo With a date stamp from the source of: 2020-09-10

References

https://www.ncib.nlm.nih.gov/entrez/query.fcgi?db=gene

See Also

AnnotationDb-class for use of the select() interface.

Examples

## select() interface:
## Objects in this package can be accessed using the select() interface
## from the AnnotationDbi package. See ?select for details.

## Bimap interface:
# Convert the object to a list
xx <- as.list(GOTERM)
if(length(xx) > 0){
    # Get the TERMS for the first elent of xx
    GOID(xx[[1]])
    Term(xx[[1]])
    Synonym(xx[[1]])
    Secondary(xx[[1]])
    Definition(xx[[1]])
    Ontology(xx[[1]])
}

[Package GO.db version 3.12.1 Index]