GO_dbconn {GO.db}	R Documentation

Collect information about the package annotation DB

Description

Some convenience functions for getting a connection object to (or collecting information about) the package annotation DB.

Usage

  GO_dbconn()
  GO_dbfile()
  GO_dbschema(file="", show.indices=FALSE)
  GO_dbInfo()

Arguments

file
A connection, or a character string naming the file to print to (see the file argument of the cat function for the details).

show.indices
The CREATE INDEX statements are not shown by default. Use show.indices=TRUE to get them.

Details

GO_dbconn returns a connection object to the package annotation DB. IMPORTANT: Don't call dbDisconnect on the connection object returned by GO_dbconn or you will break all the AnnDbObj objects defined in this package!

GO_dbfile returns the path (character string) to the package annotation DB (this is an SQLite file).

GO_dbschema prints the schema definition of the package annotation DB.

GO_dbInfo prints other information about the package annotation DB.

Value

GO_dbconn: a DBIConnection object representing an open connection to the package annotation DB.

GO_dbfile: a character string with the path to the package annotation DB.

GO_dbschema: none (invisible NULL).

GO_dbInfo: none (invisible NULL).

See Also

dbGetQuery, dbConnect, dbconn, dbfile, dbschema, dbInfo

Examples

  library(DBI)
  ## Count the number of rows in the "go_term" table:
  dbGetQuery(GO_dbconn(), "SELECT COUNT(*) FROM go_term")

  GO_dbschema()

  GO_dbInfo()

[Package GO.db version 3.12.1 Index]